third_party/blink/web_tests/fast/html/imports/import-script-prototypes-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
This test ensures each import and master is in same JS context.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS stringInMaster.__proto__ is stringInImport.__proto__
PASS thisDocument !== document is true
PASS thisDocument.__proto__ is document.__proto__
PASS thisDocument.head.__proto__ is document.head.__proto__
PASS (new Image()).ownerDocument is document
PASS thisDocument.createElement('div').ownerDocument is thisDocument
PASS document.createElement('div').ownerDocument is document
PASS successfullyParsed is true

TEST COMPLETE
